+++ dzVents/runtime/Timer.lua
@@ -2,7 +2,13 @@ local function Timer(domoticz, rule)
 
     local self = {
         baseType = domoticz.BASETYPE_TIMER,
-        triggerRule = rule
+        triggerRule = rule,
+        isVariable = false,
+		isHTTPResponse = false,
+	    isDevice = false,
+	    isScene = false,
+	    isGroup = false,
+	    isTimer = true,
     }
 
     return self